// web-app/src/app/models/user.model.ts
export interface User {
  name?: string;
  avatar?: string;
  email?: string;
  age?: number;
  login?: string;
  password?: string;
  status?: boolean;
  department?: string;
  groups?: string;
  github?: string;
  facebook?: string;
  twitter?: string;
  instagram?: string;
  description?: string;
  files?: string;
  id: string;
}